Split the schema into smaller ones in the folder schemas
Split the schema into smaller ones in the folder schemas and finally combined them into the file schema.ts
