+++ src/schema.ts
@@ -1,46 +1,15 @@
+// src/schema.ts
 import { gql } from "apollo-server";
+import { mergeTypeDefs } from '@graphql-tools/merge';
+import { customerTypeDefs } from './schemas/customerSchema';
+import { lendingRecordTypeDefs } from './schemas/lendingRecordSchema';
+import { bookTypeDefs } from './schemas/bookSchema';
+import { queryAndMutationTypeDefs } from './schemas/queryAndMutationSchema';
 
-export const typeDefs = gql`
-  type Customer {
-    id: String!
-    name: String!
-    email: String!
-  }
-
-  type LendingRecord {
-    id: String!
-    customer: Customer!
-    lentDate: String!
-    returnDate: String
-    dueDate: String!
-  }
-
-  type Book {
-    id: String!
-    title: String!
-    author: String!
-    isLent: Boolean!
-    currentLendee: Customer
-    dueDate: String
-    lendingHistory: [LendingRecord!]!
-  }
-
-  type Query {
-    books: [Book!]!
-    book(id: String!): Book
-    customers: [Customer!]!
-    customer(id: String!): Customer
-  }
-
-  type Mutation {
-    addBook(title: String!, author: String!): Book
-    addCustomer(name: String!, email: String!): Customer
-    lendBook(
-      bookId: String!
-      customerId: String!
-      lentDate: String!
-      dueDate: String!
-    ): Book
-    returnBook(bookId: String!, returnDate: String!): Book
-  }
-`;
+// Merge all the type definitions into one
+export const typeDefs = mergeTypeDefs([
+  customerTypeDefs,
+  lendingRecordTypeDefs,
+  bookTypeDefs,
+  queryAndMutationTypeDefs,
+]);
