+++ src/schemas/customerSchema.ts
@@ -1,0 +1,9 @@
+import { gql } from "apollo-server";
+
+export const customerTypeDefs = gql`
+  type Customer {
+    id: String!
+    name: String!
+    email: String!
+  }
+`;
