+++ src/schemas/lendingRecordSchema.ts
@@ -1,0 +1,11 @@
+import { gql } from "apollo-server";
+
+export const lendingRecordTypeDefs = gql`
+  type LendingRecord {
+    id: String!
+    customer: Customer!
+    lentDate: String!
+    returnDate: String
+    dueDate: String!
+  }
+`;
